runme.py much simplified
