+++ urbs/__init__.py
@@ -18,4 +18,6 @@ from .output import get_constants, get_timeseries
 from .plot import plot, result_figures, to_color
 from .pyomoio import get_entity, get_entities, list_entities
 from .report import report
+from .runfunctions import *
 from .saveload import load, save
+from .scenarios import *
